Make renderer f32 and split shared raycore crate (Stage 0+1)
Stage 0: flip Number from f64 to f32 (rust-gpu targets f32). Three book
reference tests needed a 1e-3 tolerance (cone/cylinder t-values, materials
specular) where f32 rounding exceeds EPSILON.

Stage 1: extract a `raycore` library crate holding the math + full renderer
(single source of truth for the upcoming CPU+GPU split). The `rusttracer`
binary keeps only host concerns (canvas, colors, camera, viewport, obj_parser)
and re-exports raycore's modules so existing paths resolve unchanged.

Pure f32 + code-move; behavior unchanged. 265 tests green (246 raycore + 19 bin).

+++ raycore/src/tuples.rs
@@ -3,7 +3,7 @@ use std::ops::Div;
 use std::ops::Mul;
 use std::ops::Neg;
 use std::ops::Sub;
-pub type Number = f64;
+pub type Number = f32;
 
 pub const EPSILON: Number = 1e-5;
